requirements.txt
Django==4.2.5
# uWSGI==2.0.20 #23
# django-split-settings==1.2.0
pytest-django==4.7.0
pytest-cov==4.1.0
# factory-boy==3.3.0
# pytest-factoryboy==2.6.0
Pillow==10.1.0 
djangorestframework==3.14.0
psycopg2-binary==2.9.9
django-cors-headers==4.3.1 
gunicorn==21.2.0